requirements.txt
kafka-python==2.0.2
requests==2.31.0
snowflake-connector-python==3.6.0
pytest==7.4.4
pytest-cov==4.1.0
ruff==0.3.0
